Change get... calls to latest spec         https://bugs.webkit.org/show_bug.cgi?id=30091
        Removed old versions of get calls on WebGLRenderingContext and
        added new ones per spec returning "any". New code simplifies
        GraphicsContext3D and fixes previously unimplemented routines.
        Added custom JS and V8 bindings. Added exhaustive test case
        exercising all new code paths. Updated preexisting test cases for
        new APIs. Fixed preexisting bugs in WebKit's and Chrome's WebGL
        implementations.

        Ran WebGL layout tests in WebKit (clean) and Chrome (couple of
        preexisting known failures) and manual WebGL tests in both
        browsers.



Canonical link: https://commits.webkit.org/42777@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@51348 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/html/canvas/CanvasObject.cpp
@@ -34,6 +34,7 @@ namespace WebCore {
     
 CanvasObject::CanvasObject(WebGLRenderingContext* context)
     : m_object(0)
+    , m_shouldDeleteObject(true)
     , m_context(context)
 {
 }
@@ -44,24 +45,27 @@ CanvasObject::~CanvasObject()
         m_context->removeObject(this);
 }
 
-void CanvasObject::setObject(Platform3DObject object)
+void CanvasObject::setObject(Platform3DObject object, bool shouldDeleteObject)
 {
     if (object == m_object)
         return;
         
     deleteObject();
     m_object = object;
+    m_shouldDeleteObject = shouldDeleteObject;
 }
 
 void CanvasObject::deleteObject()
 {
     if (m_object) {
-        if (m_context) {
-            m_context->graphicsContext3D()->makeContextCurrent();
-            _deleteObject(m_object);
-        }
+        if (m_shouldDeleteObject)
+            if (m_context) {
+                m_context->graphicsContext3D()->makeContextCurrent();
+                _deleteObject(m_object);
+            }
         m_object = 0;
     }
+    m_shouldDeleteObject = true;
 }
 
 }
